And let's adjust the enemy-updating from an `if`/else to a `cond`.
If the enemy has an entrance path, proceed through it until
completion; if the enemy has completed the entrance path but hasn't
traveled the straight line to their final destination; if all else has
been exhausted, travel back and forth.

We'll also used that last one to indicate that we can start the
`espas-enemy-position` counter so we don't have to worry about what
it'll be at until the very first enemy gets into position.